happy path ampc dht test and split ampc into multiple files
