+++ crates/core/src/entrypoint/webgraph_server.rs
@@ -29,12 +29,12 @@ use crate::config::WebgraphGranularity;
 use crate::distributed::cluster::Cluster;
 use crate::distributed::member::Member;
 use crate::distributed::member::Service;
+use crate::distributed::sonic::service::sonic_service;
 use crate::distributed::sonic::service::Message;
 use crate::ranking::inbound_similarity::InboundSimilarity;
 use crate::searcher::DistributedSearcher;
 use crate::searcher::SearchClient;
 use crate::similar_hosts::SimilarHostsFinder;
-use crate::sonic_service;
 use crate::webgraph::Compression;
 use crate::webgraph::FullEdge;
 use crate::webgraph::Node;
